+++ CHANGELOG.md
@@ -13,6 +13,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Deleted
 
+## [1.1.0] - 2025-05-24
+
+### Added
+
+- Add [Reddit](https://www.reddit.com/) wrapper
+
+### Changed
+
+- Environment variable names for email sender
+
+### Deleted
+
+- Remove config class
+
 ## [1.0.2] - 2025-02-28
 
 ## Changed